chore: wave 2 architectural hardening complete
