+++ apps/mobile/babel.config.js
@@ -21,6 +21,12 @@ module.exports = function (api) {
             '@protocols': './src/features/protocols',
             '@stock': './src/features/stock',
             '@dashboard': './src/features/dashboard',
+            '@navigation': './src/navigation',
+            '@platform': './src/platform',
+            '@profile': './src/features/profile',
+            '@notifications': './src/features/notifications',
+            '@treatments': './src/features/treatments',
+            '@dose': './src/features/dose',
           },
         },
       ],
