+++ apps/mobile/src/features/dashboard/screens/TodayScreen.jsx
@@ -10,21 +10,21 @@ import {
   UIManager 
 } from 'react-native'
 import { Pill } from 'lucide-react-native'
-import { useTodayData } from '../hooks/useTodayData'
-import ScreenContainer from '../../../shared/components/ui/ScreenContainer'
-import LoadingState from '../../../shared/components/states/LoadingState'
-import EmptyState from '../../../shared/components/states/EmptyState'
-import ErrorState from '../../../shared/components/states/ErrorState'
+import { useTodayData } from '@dashboard/hooks/useTodayData'
+import ScreenContainer from '@shared/components/ui/ScreenContainer'
+import LoadingState from '@shared/components/states/LoadingState'
+import EmptyState from '@shared/components/states/EmptyState'
+import ErrorState from '@shared/components/states/ErrorState'
 import { getPeriodFromTime, getNow } from '@dosiq/core'
-import AdherenceDayCard from '../components/AdherenceDayCard'
-import TimeBlockSeparator from '../components/TimeBlockSeparator'
-import DoseTimelineCard from '../components/DoseTimelineCard'
-import HeroDoseCard from '../components/HeroDoseCard'
-import StockAlertInline from '../components/StockAlertInline'
-import DoseRegisterModal from '../../dose/components/DoseRegisterModal'
-import BulkDoseRegisterModal from '../../dose/components/BulkDoseRegisterModal'
-import StaleBanner from '../../../shared/components/feedback/StaleBanner'
-import { colors, spacing, typography } from '../../../shared/styles/tokens'
+import AdherenceDayCard from '@dashboard/components/AdherenceDayCard'
+import TimeBlockSeparator from '@dashboard/components/TimeBlockSeparator'
+import DoseTimelineCard from '@dashboard/components/DoseTimelineCard'
+import HeroDoseCard from '@dashboard/components/HeroDoseCard'
+import StockAlertInline from '@dashboard/components/StockAlertInline'
+import DoseRegisterModal from '@dose/components/DoseRegisterModal'
+import BulkDoseRegisterModal from '@dose/components/BulkDoseRegisterModal'
+import StaleBanner from '@shared/components/feedback/StaleBanner'
+import { colors, spacing, typography } from '@shared/styles/tokens'
 
 // Habilitar animações no Android
 if (Platform.OS === 'android' && UIManager.setLayoutAnimationEnabledExperimental) {
