+++ apps/mobile/src/platform/auth/authService.js
@@ -1,6 +1,6 @@
 // authService.js — serviço de autenticação com validação Zod
 import { z } from 'zod'
-import { supabase } from '../supabase/nativeSupabaseClient'
+import { supabase } from '@platform/supabase/nativeSupabaseClient'
 
 /**
  * Schema para validação de credenciais de login
